backend/requirements.txt
django>=4.2,<5
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
requests>=2.31,<3
waitress>=2.1,<3

